requirements.txt
# =============================================================================
# Backend Dependencies only packages actually imported in backend/ code
# =============================================================================

--extra-index-url https://download.pytorch.org/whl/cu121
--find-links https://download.pytorch.org/whl/torch_stable.html

# --- Web Framework & Server ---
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic==2.11.9

# --- HTTP Clients ---
requests==2.32.3
httpx==0.28.1
aiofiles==24.1.0

# --- Audio: Generation & Processing ---
# Mac will fetch standard wheels, Windows will fetch cu121 wheels
torch==2.4.0+cu121 ; sys_platform != 'darwin'
torchaudio==2.4.0+cu121 ; sys_platform != 'darwin'
torchvision==0.19.0+cu121 ; sys_platform != 'darwin'
torch==2.4.0 ; sys_platform == 'darwin'
torchaudio==2.4.0 ; sys_platform == 'darwin'
torchvision==0.19.0 ; sys_platform == 'darwin'


diffusers==0.30.0
safetensors==0.4.4
transformers==4.44.2
huggingface-hub==0.24.6
soundfile==0.13.1
librosa==0.11.0
noisereduce==3.0.3
tensorflow==2.18.1
tensorflow-hub==0.16.1

# --- Room Acoustics ---
pyroomacoustics==0.9.0
scipy==1.15.3
numpy==1.26.4
matplotlib==3.10.7

# --- Speckle (3D model platform) ---
specklepy==3.2.3
gql==3.5.3

# --- Misc ---
numpy-stl==3.2.0
pythonnet==3.0.5
